requirements.txt
Fabric>=1.6.0
servers.py>=0.1.4